Providing grid_ggd and space info thru Dict/YAML
populate_grid_ggd has a new first argument to input information about grid_ggd and space as a dictionary or as a string which is filename of a yaml file containing the information. A sample file ./sample/gridspacedesc.yml has been created for running the test which keeps same information as what David used for playing around earlier.