Added x points node in subset and made code leaner
x point in the grid is found using the cuts information and stored
in a corresponding grid_subset. In an effort to make the code leaner,
all assignments to grid_subset elements are now handled through
common function add_subset_element! which adds the fields in place
and appends to the elements array. Thus there is no need to store
and manage element array indices separately. I don't think there is
a huge penalty in running the code if we resize the array for each
element, it seems like it can take it. This also ensures optimized
size of the element array and there are no empty elements in a subset.

find_subset_index function is kept but its use has been moved to
get_grid_subset_with_index function that directly returns the grid_subset
object with the matching index and does not rely on gsdesc, instead uses the
datastructure itself. This will also help in finding any errors in code in
future. Note, that a similar search and return function can be made using the
grid_subset name for an easy to read code. However, I think grid_subset indices
are the fundamental identifiers and not the names.