Test to see issue in imas2json
This simple test shows that the imas2json function is not working
properly. The test is run with the following command:

```julia
julia --project=. test/runtests.jl
```

Note: for running this test, you will need to run `dvc pull` to get the
sample files.

There is some issue in writing down the data inside electrons.density,
electrons.temperature and so on. You can check the json file generated
by a json viewer as well.