Adding ggd values using setproperty!
OMAS.jl/data.jl overloads Base.keys() such that it only returns the
fields in the data structure whose symbol names are stored in an
internal variable called _filled. This list is updated by the overloaded
setproperty! function which is called when a value is assigned to a
field using explicit "=" or use of "setproperty!". We on the other hand
were using resize to create an array for data fields in ggd and filling
them in using the index. This does not update the _filled list and
thus keys() does not return the new fields. Because of this, imas2dict
function doesnot copy the new fields to the output dictionary and thus
imas2json function was failing for our use case. See issue
https://github.com/ProjectTorreyPines/OMAS.jl/issues/9

This commit fixes this by modifying the val_obj function to return
a parent and property symbol instead of the property object directly.
The value is assigned to the property using setproperty! which then
correctly updates teh internal list _filled. This fixes the issue.