Adding new test files ofr testing parser using DVC
Removed:
	deleted:    samples/b2mn.dat.sample_50dn
	deleted:    samples/b2mn.dat.sample_50xd
	deleted:    samples/b2mn.dat.sample_si1
	deleted:    samples/b2mn.dat.sample_si2

Added:
    new file:   samples/b2mn.dat.json.dvc
    new file:   samples/test_b2mn.dat.dvc
    new file:   samples/test_b2mn.dat.json.dvc

Adding a json file of correctly parsed version of b2mn.dat file that is
already in the samples directory. Also added a new test_b2mn.dat file
that is manually edited to cover all edge cases of the parser and a
json file version of the same that has been correctly parsed (checked
manually).