Using white list for integer fields and some default values
Created a new file in src/b2mn_int_fields.txt where names of all known
int fields are listed. This file is used to check if a field value
should be converted to integer. Also, if there is a second value listed
in this file in a row, that field is considered to be required
and a default value is used if it is not found in the input file.

Updated sample json files with additional int fields that get added
by default.