Ensuring ggd values refer to correct grid_subset
When EIRENE grid is added, the b2.5 grid is moved to negative
grid_subset_index. For example, all cells of b2.5 grid are now present
in grid_susbset with index -5. Thus, any quantities that are defined on
cells of b2.5 should refer ot -5 instead of 5.

This is implemented by providing an additional argument to the val_obj
function which is adictionary of grid_subset_index changes. If EIRENE
grid is added, we populate this dictionary to show that 1 => -1, 2 => -2
and 5 => -5. This way, the ggd values are correctly set.

Now the comment and hotfix in https://github.com/ProjectTorreyPines/SynthDiag.jl/blob/806f1beffa8e72948b36c821ca2d628c8cb3174f/src/interferometer.jl#L68
can be removed if we'll rely on this extended grid and not the one
that is generated in SD4SOLPS.