Quick bug fix: Allow empty elements in fort.33
When fnnodes is not a multiple of 4, therea res ome rows in fort.33 that
correspond to the last rowf or fnodeX and fnodeY that may contain
nothing. When julia reads a file linke this, it creates an empty string
when nothing is present at a matrix location. This should be handled
gracefully. This commit fixes this issue. It converts empty string
into missing and then skips missing after vectorizing the matrix.