Quick bug fix: Read scale parameters for b2time.nc
The scale parameters for the b2time.nc file were not being read. This
created a unit mismatch for temperature of ions and electrons as they
were stored in J and IMAS requires eV. The NetCDF dataset however
also comes alone with an attribute scale whenever the quantity needs to
be converted in units. This commit adds the reading of the scale
attribute and applies it whenever found. This fixes the reading of
temperature data.