Need to jest with Julia >= 1.10
IMASDD depends on DataInterpolations >=5.0
Latest version of DataInterpolations is 5.2.0 depends on LinearAlgebra >= 1.10
LinearAlgebra is Julia's standard library, so our code now depends on Julia >= 1.10

Like IMASDD we should also test and build our repo for latest subversion of Julia 1.x
This commit makes that change. For anyone testing this locally, you will have to upgrade to Julia 1.10 or higher.
Best way to do this is to use the Juliaup tool, which is a Julia package that can be installed with the following command:
curl -fsSL https://install.julialang.org | sh

See https://julialang.org/downloads/ for more information.