Temporarily removing Fortran90Namelists dep
Fortran90Namelists is not registered package with Julia and is being used
for only one function in this repo right now:

read_b2_boundary_parameters

This commit comments out any usage of Fortran90Namelists and removes it
from dependency list. Once this package is registered with Julia, this
feature will be brought back in.