Adding back read_b2_boundary_parameters
This function was removed in 0d2ef6bbc031f67875aaece81607481c64af681d
to remove unregistered Fortran90Namelists dep. The usecase of reading
the namelist was very specific, so this commit adds a small basic julia
funciton to read the required parts of namelist and adds back this function.

In future, if generic fortran namelist parser is available in Julia Registry,
we will use that.